+++ Realm/Realm/Configurations/FlexibleSyncConfiguration.cs
@@ -18,6 +18,8 @@
 
 using System;
 using System.Diagnostics.CodeAnalysis;
+using System.Threading;
+using System.Threading.Tasks;
 using Realms.Sync.ErrorHandling;
 
 namespace Realms.Sync
@@ -32,6 +34,12 @@ namespace Realms.Sync
     public class FlexibleSyncConfiguration : SyncConfigurationBase
     {
         /// <summary>
+        /// A delegate invoked when a flexible sync Realm is first opened.
+        /// </summary>
+        /// <param name="realm">The realm that has just been opened.</param>
+        public delegate void InitialSubscriptionsDelegate(Realm realm);
+
+        /// <summary>
         /// Initializes a new instance of the <see cref="FlexibleSyncConfiguration"/> class.
         /// </summary>
         /// <param name="user">
@@ -45,6 +53,117 @@ namespace Realms.Sync
             : base(user)
         {
             DatabasePath = GetPathToRealm(optionalPath ?? user.App.Handle.GetRealmPath(User));
+        }
+
+        /// <summary>
+        /// Gets or sets a callback that will be invoked the first time a Realm is opened.
+        /// </summary>
+        /// <remarks>
+        /// This callback allows you to populate an initial set of subscriptions, which will
+        /// then be awaited when <see cref="Realm.GetInstance(RealmConfigurationBase)"/> is invoked.
+        /// <br/>
+        /// The <see cref="SubscriptionSet"/> returned by <see cref="Realm.Subscriptions"/> is already
+        /// called in an <see cref="SubscriptionSet.Update(System.Action)"/> block, so there's no need
+        /// to start one inside the callback.
+        /// </remarks>
+        /// <example>
+        /// <code>
+        /// var config = new FlexibleSyncConfiguration(user)
+        /// {
+        ///     PopulateInitialSubscriptions = (realm) =>
+        ///     {
+        ///         var myNotes = realm.All&lt;Note&gt;().Where(n =&gt; n.AuthorId == myUserId);
+        ///         realm.Subscriptions.Add(myNotes);
+        ///     }
+        /// };
+        ///
+        /// // The task will complete when all the user notes have been downloaded.
+        /// var realm = await Realm.GetInstanceAsync(config);
+        /// </code>
+        /// </example>
+        /// <value>
+        /// The <see cref="InitialSubscriptionsDelegate"/> that will be invoked the first time
+        /// a Realm is opened.
+        /// </value>
+        public InitialSubscriptionsDelegate PopulateInitialSubscriptions { get; set; }
+
+        internal override async Task<Realm> CreateRealmAsync(CancellationToken cancellationToken)
+        {
+            var tracker = InjectInitialSubscriptions();
+
+            var result = await base.CreateRealmAsync(cancellationToken);
+
+            InvokeInitialSubscriptions(tracker, result);
+
+            if (tracker.PopulateInitialDataInvoked)
+            {
+                await result.Subscriptions.WaitForSynchronizationAsync();
+            }
+
+            return result;
+        }
+
+        internal override Realm CreateRealm()
+        {
+            var tracker = InjectInitialSubscriptions();
+
+            var result = base.CreateRealm();
+
+            InvokeInitialSubscriptions(tracker, result);
+
+            return result;
+        }
+
+        private InitialDataTracker InjectInitialSubscriptions()
+        {
+            var tracker = new InitialDataTracker
+            {
+                Callback = PopulateInitialSubscriptions
+            };
+
+            if (tracker.Callback != null)
+            {
+                var oldDataCallback = PopulateInitialData;
+                PopulateInitialData = (realm) =>
+                {
+                    // We can't run the PopulateInitialSubscriptions callback here because
+                    // the Realm is already in a write transaction, so Subscriptions.Update
+                    // will hang. We flag that with `shouldPopulate` and update the subscriptions
+                    // after we open the realm.
+                    oldDataCallback?.Invoke(realm);
+                    tracker.PopulateInitialDataInvoked = true;
+                };
+            }
+
+            return tracker;
+        }
+
+        private static void InvokeInitialSubscriptions(InitialDataTracker tracker, Realm realm)
+        {
+            if (!tracker.PopulateInitialDataInvoked)
+            {
+                return;
+            }
+
+            try
+            {
+                realm.Subscriptions.Update(() =>
+                {
+                    // We're not guarded by a write lock between invoking the `InitialDataCallback`
+                    // and getting here, so it's possible someone managed to insert subscriptions
+                    // before us. If that's the case, then don't insert anything and just wait for
+                    // sync.
+                    if (realm.Subscriptions.Count == 0)
+                    {
+                        tracker.Callback(realm);
+                    }
+                });
+            }
+            catch (Exception ex)
+            {
+                // This needs to duplicate the logic in RealmConfigurationBase.CreateRealmAsync
+                throw new AggregateException("Exception occurred in a Realm.PopulateInitialSubscriptions callback. See inner exception for more details.", ex);
+            }
         }
 
         internal override Native.SyncConfiguration CreateNativeSyncConfiguration()
@@ -78,6 +197,16 @@ namespace Realms.Sync
 
                 base.ClientResetHandler = value;
             }
+        }
+
+        // This is a holder class to workaround the fact that we can't use ref booleans
+        // inside lambda functions. Its purpose is to server as a marker that PopulateInitialData
+        // was invoked, which means we need to then update subscriptions.
+        private class InitialDataTracker
+        {
+            public bool PopulateInitialDataInvoked;
+
+            public InitialSubscriptionsDelegate Callback;
         }
     }
 }