+++ Realm/Realm/Exceptions/RealmException.cs
@@ -98,9 +98,6 @@ namespace Realms.Exceptions
                 case RealmExceptionCodes.RealmDuplicatePrimaryKeyValue:
                     return new RealmDuplicatePrimaryKeyValueException(message);
 
-                case RealmExceptionCodes.RealmDotNetExceptionDuringMigration:
-                    return new ManagedExceptionDuringMigrationException(message);
-
                 case RealmExceptionCodes.RealmClosed:
                     return new RealmClosedException(message);