+++ Realm/Realm/Handles/SharedRealmHandle.cs
@@ -28,7 +28,6 @@ using Realms.Logging;
 using Realms.Native;
 using Realms.Schema;
 using Realms.Sync;
-using static Realms.RealmConfiguration;
 
 namespace Realms
 {
@@ -73,13 +72,14 @@ namespace Realms
 
             // migrationSchema is a special schema that is used only in the context of a migration block.
             // It is a pointer because we need to be able to modify this schema in some migration methods directly in core.
-            [return: MarshalAs(UnmanagedType.U1)]
             [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-            internal delegate bool MigrationCallback(IntPtr oldRealm, IntPtr newRealm, IntPtr migrationSchema, Native.Schema oldSchema, ulong schemaVersion, IntPtr managedMigrationHandle);
+            internal delegate IntPtr MigrationCallback(IntPtr oldRealm, IntPtr newRealm, IntPtr migrationSchema, Native.Schema oldSchema, ulong schemaVersion, IntPtr managedMigrationHandle);
 
-            [return: MarshalAs(UnmanagedType.U1)]
             [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-            internal delegate bool ShouldCompactCallback(IntPtr config, ulong totalSize, ulong dataSize);
+            internal delegate IntPtr ShouldCompactCallback(IntPtr managedDelegate, ulong totalSize, ulong dataSize, [MarshalAs(UnmanagedType.U1)] ref bool should_compact);
+
+            [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
+            internal delegate IntPtr InitializationCallback(IntPtr managedInitializationDelegate, IntPtr realm);
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_open", CallingConvention = CallingConvention.Cdecl)]
             public static extern IntPtr open(Configuration configuration,
@@ -192,7 +192,8 @@ namespace Realms
                 DictionaryHandle.KeyNotificationCallback notify_dictionary,
                 MigrationCallback migration_callback,
                 ShouldCompactCallback should_compact_callback,
-                HandleTaskCompletionCallback handle_task_completion);
+                HandleTaskCompletionCallback handle_task_completion,
+                InitializationCallback initialization_callback);
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_has_changed", CallingConvention = CallingConvention.Cdecl)]
             [return: MarshalAs(UnmanagedType.U1)]
@@ -252,6 +253,7 @@ namespace Realms
             NativeMethods.MigrationCallback onMigration = OnMigration;
             NativeMethods.ShouldCompactCallback shouldCompact = ShouldCompactOnLaunchCallback;
             NativeMethods.HandleTaskCompletionCallback handleTaskCompletion = HandleTaskCompletionCallback;
+            NativeMethods.InitializationCallback onInitialization = OnDataInitialization;
 
             GCHandle.Alloc(notifyRealm);
             GCHandle.Alloc(getNativeSchema);
@@ -263,8 +265,9 @@ namespace Realms
             GCHandle.Alloc(onMigration);
             GCHandle.Alloc(shouldCompact);
             GCHandle.Alloc(handleTaskCompletion);
+            GCHandle.Alloc(onInitialization);
 
-            NativeMethods.install_callbacks(notifyRealm, getNativeSchema, openRealm, disposeGCHandle, logMessage, notifyObject, notifyDictionary, onMigration, shouldCompact, handleTaskCompletion);
+            NativeMethods.install_callbacks(notifyRealm, getNativeSchema, openRealm, disposeGCHandle, logMessage, notifyObject, notifyDictionary, onMigration, shouldCompact, handleTaskCompletion, onInitialization);
         }
 
         [Preserve]
@@ -403,10 +406,10 @@ namespace Realms
             return new SharedRealmHandle(result);
         }
 
-        public static AsyncOpenTaskHandle OpenWithSyncAsync(Configuration configuration, Sync.Native.SyncConfiguration syncConfiguration, RealmSchema schema, byte[] encryptionKey, GCHandle tcsHandle)
+        public static AsyncOpenTaskHandle OpenWithSyncAsync(Configuration configuration, Sync.Native.SyncConfiguration syncConfiguration, RealmSchema schema, byte[] encryptionKey, IntPtr tcsHandle)
         {
             var marshaledSchema = new SchemaMarshaler(schema);
-            var asyncTaskPtr = NativeMethods.open_with_sync_async(configuration, syncConfiguration, marshaledSchema.Objects, marshaledSchema.Objects.Length, marshaledSchema.Properties, encryptionKey, GCHandle.ToIntPtr(tcsHandle), out var nativeException);
+            var asyncTaskPtr = NativeMethods.open_with_sync_async(configuration, syncConfiguration, marshaledSchema.Objects, marshaledSchema.Objects.Length, marshaledSchema.Properties, encryptionKey, tcsHandle, out var nativeException);
             nativeException.ThrowIfNecessary();
             return new AsyncOpenTaskHandle(asyncTaskPtr);
         }
@@ -592,7 +595,9 @@ namespace Realms
         {
             var useSync = config is SyncConfigurationBase;
 
-            NativeMethods.write_copy(this, config.CreateNativeConfiguration(), useSync, config.EncryptionKey, out var nativeException);
+            var nativeConfig = config.CreateNativeConfiguration();
+
+            NativeMethods.write_copy(this, nativeConfig, useSync, config.EncryptionKey, out var nativeException);
             nativeException.ThrowIfNecessary();
         }
 
@@ -752,34 +757,57 @@ namespace Realms
         }
 
         [MonoPInvokeCallback(typeof(NativeMethods.MigrationCallback))]
-        private static bool OnMigration(IntPtr oldRealmPtr, IntPtr newRealmPtr, IntPtr migrationSchema, Native.Schema oldSchema, ulong schemaVersion, IntPtr managedMigrationHandle)
+        private static IntPtr OnMigration(IntPtr oldRealmPtr, IntPtr newRealmPtr, IntPtr migrationSchema, Native.Schema oldSchema, ulong schemaVersion, IntPtr managedConfigHandle)
         {
-            var migrationHandle = GCHandle.FromIntPtr(managedMigrationHandle);
-            var migration = (Migration)migrationHandle.Target;
-
-            var oldRealmHandle = new UnownedRealmHandle(oldRealmPtr);
-            var oldConfiguration = new RealmConfiguration(migration.Configuration.DatabasePath)
+            Migration migration = null;
+            try
             {
-                SchemaVersion = schemaVersion,
-                IsReadOnly = true,
-                EnableCache = false
-            };
-            var oldRealm = new Realm(oldRealmHandle, oldConfiguration, RealmSchema.CreateFromObjectStoreSchema(oldSchema));
+                var configHandle = GCHandle.FromIntPtr(managedConfigHandle);
+                var config = (RealmConfiguration)configHandle.Target;
 
-            var newRealmHandle = new UnownedRealmHandle(newRealmPtr);
-            var newRealm = new Realm(newRealmHandle, migration.Configuration, migration.Schema, isInMigration: true);
+                var oldRealmHandle = new UnownedRealmHandle(oldRealmPtr);
+                var oldConfiguration = new RealmConfiguration(config.DatabasePath)
+                {
+                    SchemaVersion = schemaVersion,
+                    IsReadOnly = true,
+                    EnableCache = false
+                };
+                using var oldRealm = new Realm(oldRealmHandle, oldConfiguration, RealmSchema.CreateFromObjectStoreSchema(oldSchema));
 
-            var result = migration.Execute(oldRealm, newRealm, migrationSchema);
+                var newRealmHandle = new UnownedRealmHandle(newRealmPtr);
+                using var newRealm = new Realm(newRealmHandle, config, config.GetSchema(), isInMigration: true);
+                migration = new Migration(oldRealm, newRealm, migrationSchema);
 
-            return result;
+                config.MigrationCallback.Invoke(migration, schemaVersion);
+                return IntPtr.Zero;
+            }
+            catch (Exception ex)
+            {
+                var exHandle = GCHandle.Alloc(ex);
+                return GCHandle.ToIntPtr(exHandle);
+            }
+            finally
+            {
+                migration?.Free();
+            }
         }
 
         [MonoPInvokeCallback(typeof(NativeMethods.ShouldCompactCallback))]
-        private static bool ShouldCompactOnLaunchCallback(IntPtr delegatePtr, ulong totalSize, ulong dataSize)
+        private static IntPtr ShouldCompactOnLaunchCallback(IntPtr managedConfigHandle, ulong totalSize, ulong dataSize, ref bool shouldCompact)
         {
-            var handle = GCHandle.FromIntPtr(delegatePtr);
-            var compactDelegate = (ShouldCompactDelegate)handle.Target;
-            return compactDelegate(totalSize, dataSize);
+            try
+            {
+                var configHandle = GCHandle.FromIntPtr(managedConfigHandle);
+                var config = (RealmConfiguration)configHandle.Target;
+
+                shouldCompact = config.ShouldCompactOnLaunch.Invoke(totalSize, dataSize);
+                return IntPtr.Zero;
+            }
+            catch (Exception ex)
+            {
+                var exHandle = GCHandle.Alloc(ex);
+                return GCHandle.ToIntPtr(exHandle);
+            }
         }
 
         [MonoPInvokeCallback(typeof(NativeMethods.HandleTaskCompletionCallback))]
@@ -829,6 +857,26 @@ namespace Realms
                     tcs.TrySetCanceled();
                 }
             }, null);
+        }
+
+        [MonoPInvokeCallback(typeof(NativeMethods.InitializationCallback))]
+        private static IntPtr OnDataInitialization(IntPtr managedConfigHandle, IntPtr realmPtr)
+        {
+            try
+            {
+                var configHandle = GCHandle.FromIntPtr(managedConfigHandle);
+                var config = (RealmConfigurationBase)configHandle.Target;
+
+                var realmHandle = new UnownedRealmHandle(realmPtr);
+                using var realm = config.GetRealm(realmHandle);
+                config.PopulateInitialData.Invoke(realm);
+                return IntPtr.Zero;
+            }
+            catch (Exception ex)
+            {
+                var exHandle = GCHandle.Alloc(ex);
+                return GCHandle.ToIntPtr(exHandle);
+            }
         }
 
         public class SchemaMarshaler